arrow key jadi aswd
