--- uas_pti/src/arrow_key.css
@@ -1,290 +1,0 @@
-/* arrow_key.css */
-@import url("https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap");
-
-:root {
-  --arrow-primary: #79d959;
-  --arrow-secondary: #9be686;
-  --arrow-hover: #8ae566;
-  --arrow-active: #6bc951;
-  --arrow-shadow: rgba(121, 217, 89, 0.3);
-  --pixel-size: 3px;
-  --border-radius: 12px;
-}
-
-.arrow-key-container {
-  position: fixed;
-  bottom: 160px;
-  right: 150px;
-  z-index: 50;
-}
-
-.arrow-key-pad {
-  display: grid;
-  grid-template-rows: auto auto auto;
-  gap: 8px;
-  background: #ffffff;
-  border: var(--pixel-size) solid var(--arrow-primary);
-  border-radius: var(--border-radius);
-  padding: 16px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.12);
-  image-rendering: pixelated;
-}
-
-.arrow-key-middle-row {
-  display: grid;
-  grid-template-columns: auto auto auto;
-  gap: 8px;
-  align-items: center;
-}
-
-.arrow-key {
-  width: 48px;
-  height: 48px;
-  background: linear-gradient(135deg, #f8f8ff 0%, #f0f8f0 100%);
-  border: 2px solid var(--arrow-secondary);
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.2s cubic-bezier(0.175, 0.885, 0.32, 1.275);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.08), inset 0 1px 0 rgba(255, 255, 255, 0.8);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-family: "Press Start 2P", system-ui;
-  margin: 0 auto;
-  position: relative;
-  overflow: hidden;
-}
-
-.arrow-key::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 100%;
-  background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.4), transparent);
-  transition: left 0.6s ease;
-}
-
-.arrow-key:hover::before {
-  left: 100%;
-}
-
-.arrow-key:hover {
-  transform: translateY(-2px) scale(1.05);
-  background: linear-gradient(135deg, var(--arrow-hover) 0%, var(--arrow-primary) 100%);
-  border-color: var(--arrow-primary);
-  box-shadow: 0 4px 8px var(--arrow-shadow), 0 2px 4px rgba(0, 0, 0, 0.1), inset 0 1px 0 rgba(255, 255, 255, 0.3);
-}
-
-.arrow-key:active {
-  transform: translateY(0) scale(0.98);
-  background: linear-gradient(135deg, var(--arrow-active) 0%, #5bb849 100%);
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1), inset 0 1px 2px rgba(0, 0, 0, 0.1);
-}
-
-.arrow-icon {
-  width: 20px;
-  height: 20px;
-  position: relative;
-  image-rendering: pixelated;
-}
-
-/* Arrow shapes using CSS */
-.arrow-icon-up::before,
-.arrow-icon-down::before,
-.arrow-icon-left::before,
-.arrow-icon-right::before {
-  content: "";
-  position: absolute;
-  border-style: solid;
-  transition: border-color 0.2s ease;
-}
-
-.arrow-icon-up::before {
-  border-width: 0 10px 16px 10px;
-  border-color: transparent transparent #4d4d4d transparent;
-  top: 2px;
-  left: 0;
-}
-
-.arrow-icon-down::before {
-  border-width: 16px 10px 0 10px;
-  border-color: #4d4d4d transparent transparent transparent;
-  top: 2px;
-  left: 0;
-}
-
-.arrow-icon-left::before {
-  border-width: 10px 16px 10px 0;
-  border-color: transparent #4d4d4d transparent transparent;
-  top: 0;
-  left: 2px;
-}
-
-.arrow-icon-right::before {
-  border-width: 10px 0 10px 16px;
-  border-color: transparent transparent transparent #4d4d4d;
-  top: 0;
-  left: 2px;
-}
-
-/* Hover states for arrows */
-.arrow-key:hover .arrow-icon-up::before {
-  border-bottom-color: white;
-}
-
-.arrow-key:hover .arrow-icon-down::before {
-  border-top-color: white;
-}
-
-.arrow-key:hover .arrow-icon-left::before {
-  border-right-color: white;
-}
-
-.arrow-key:hover .arrow-icon-right::before {
-  border-left-color: white;
-}
-
-/* Center decoration */
-.arrow-key-center {
-  width: 48px;
-  height: 48px;
-  background: #f8f8ff;
-  border: 2px solid #d8e1c5;
-  border-radius: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-shadow: inset 0 2px 4px rgba(0, 0, 0, 0.05);
-}
-
-.center-dot {
-  width: 12px;
-  height: 12px;
-  background: var(--arrow-primary);
-  border-radius: 50%;
-  box-shadow: 0 2px 4px var(--arrow-shadow);
-  position: relative;
-  overflow: hidden;
-}
-
-.center-dot::after {
-  content: "";
-  position: absolute;
-  top: 2px;
-  left: 2px;
-  width: 4px;
-  height: 4px;
-  background: rgba(255, 255, 255, 0.6);
-  border-radius: 50%;
-}
-
-/* Active key animation */
-@keyframes keyPress {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(0.95);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-.arrow-key:active .arrow-icon {
-  animation: keyPress 0.2s ease;
-}
-
-/* Pulse animation for center dot */
-@keyframes centerPulse {
-  0% {
-    box-shadow: 0 2px 4px var(--arrow-shadow), 0 0 0 0 var(--arrow-shadow);
-  }
-  70% {
-    box-shadow: 0 2px 4px var(--arrow-shadow), 0 0 0 8px rgba(121, 217, 89, 0);
-  }
-  100% {
-    box-shadow: 0 2px 4px var(--arrow-shadow), 0 0 0 0 rgba(121, 217, 89, 0);
-  }
-}
-
-.center-dot {
-  animation: centerPulse 2s infinite;
-}
-
-/* Mobile responsiveness */
-@media (max-width: 768px) {
-  .arrow-key-container {
-    bottom: 20px;
-    right: 20px;
-  }
-
-  .arrow-key-pad {
-    padding: 12px;
-    gap: 6px;
-  }
-
-  .arrow-key,
-  .arrow-key-center {
-    width: 40px;
-    height: 40px;
-  }
-
-  .arrow-key-middle-row {
-    gap: 6px;
-  }
-
-  .arrow-icon {
-    width: 16px;
-    height: 16px;
-  }
-
-  .arrow-icon-up::before,
-  .arrow-icon-down::before {
-    border-width: 0 8px 14px 8px;
-  }
-
-  .arrow-icon-down::before {
-    border-width: 14px 8px 0 8px;
-  }
-
-  .arrow-icon-left::before {
-    border-width: 8px 14px 8px 0;
-  }
-
-  .arrow-icon-right::before {
-    border-width: 8px 0 8px 14px;
-  }
-
-  .center-dot {
-    width: 10px;
-    height: 10px;
-  }
-}
-
-/* Accessibility */
-.arrow-key:focus {
-  outline: none;
-  border-color: var(--arrow-primary);
-  box-shadow: 0 4px 8px var(--arrow-shadow), 0 0 0 3px rgba(121, 217, 89, 0.4);
-}
-
-.arrow-key:focus:not(:focus-visible) {
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.08), inset 0 1px 0 rgba(255, 255, 255, 0.8);
-}
-
-/* Optional: Add keyboard indicator */
-.arrow-key-container::before {
-  content: "Arrow Keys";
-  position: absolute;
-  top: -25px;
-  left: 50%;
-  transform: translateX(-50%);
-  font-family: "Press Start 2P", system-ui;
-  font-size: 9px;
-  color: var(--arrow-primary);
-  text-shadow: 1px 1px 0px rgba(255, 255, 255, 0.8);
-  white-space: nowrap;
-}
